backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiohttp==3.9.1
neo4j==5.14.0
redis==5.0.1
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.4